Replace `CGFloat.pi` by `Double.pi`
`CGFloat.pi` doesn’t always work

+++ NVActivityIndicatorView/Animations/NVActivityIndicatorAnimationBallRotateChase.swift
@@ -61,7 +61,7 @@ class NVActivityIndicatorAnimationBallRotateChase: NVActivityIndicatorAnimationD
         let positionAnimation = CAKeyframeAnimation(keyPath: "position")
         positionAnimation.duration = duration
         positionAnimation.repeatCount = HUGE
-        positionAnimation.path = UIBezierPath(arcCenter: CGPoint(x: x, y: y), radius: size.width / 2, startAngle: 3 * CGFloat.pi * 0.5, endAngle: 3 * CGFloat.pi * 0.5 + 2 * CGFloat.pi, clockwise: true).cgPath
+        positionAnimation.path = UIBezierPath(arcCenter: CGPoint(x: x, y: y), radius: size.width / 2, startAngle: CGFloat(3 * Double.pi * 0.5), endAngle: CGFloat(3 * Double.pi * 0.5 + 2 * Double.pi), clockwise: true).cgPath
 
         // Aniamtion
         let animation = CAAnimationGroup()
