+++ NVActivityIndicatorView/Animations/NVActivityIndicatorAnimationBallSpinFadeLoader.swift
@@ -63,7 +63,7 @@ class NVActivityIndicatorAnimationBallSpinFadeLoader: NVActivityIndicatorAnimati
 
         // Draw circles
         for i in 0 ..< 8 {
-            let circle = circleAt(angle: CGFloat.pi / 4 * CGFloat(i),
+            let circle = circleAt(angle: CGFloat(Double.pi / 4) * CGFloat(i),
                                   size: circleSize,
                                   origin: CGPoint(x: x, y: y),
                                   containerSize: size,
