+++ NVActivityIndicatorView/Animations/NVActivityIndicatorAnimationTriangleSkewSpin.swift
@@ -42,9 +42,9 @@ class NVActivityIndicatorAnimationTriangleSkewSpin: NVActivityIndicatorAnimation
         animation.timingFunctions = [timingFunction, timingFunction, timingFunction, timingFunction]
         animation.values = [
             NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: 0), createRotateYTransform(angle: 0))),
-            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: CGFloat.pi), createRotateYTransform(angle: 0))),
-            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: CGFloat.pi), createRotateYTransform(angle: CGFloat.pi))),
-            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: 0), createRotateYTransform(angle: CGFloat.pi))),
+            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: CGFloat(Double.pi)), createRotateYTransform(angle: 0))),
+            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: CGFloat(Double.pi)), createRotateYTransform(angle: CGFloat(Double.pi)))),
+            NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: 0), createRotateYTransform(angle: CGFloat(Double.pi)))),
             NSValue(caTransform3D: CATransform3DConcat(createRotateXTransform(angle: 0), createRotateYTransform(angle: 0))),
         ]
         animation.duration = duration
